Correct casting for mac
git-svn-id: file:///home/jj/hercules.svn/trunk@6498 956126f8-22a0-4046-8f4a-272fa8102e63

+++ cckddiag.c
@@ -504,7 +504,7 @@ char            pathname[MAX_PATH];     /* file path in host format  */
     readpos(fd, &devhdr, 0, sizeof(devhdr));
     if (cmd_devhdr) {
         fprintf(stderr, "\nDEVHDR - %d (decimal) bytes:\n", 
-                (ssize_t)sizeof(devhdr));
+                (int)sizeof(devhdr));
         data_dump(&devhdr, sizeof(devhdr));
     }
 
@@ -571,7 +571,7 @@ char            pathname[MAX_PATH];     /* file path in host format  */
     readpos(fd, &cdevhdr, CKDDASD_DEVHDR_SIZE, sizeof(cdevhdr));
     if (cmd_cdevhdr) {
         fprintf(stderr, "\nCDEVHDR - %d (decimal) bytes:\n",
-                (ssize_t)sizeof(cdevhdr));
+                (int)sizeof(cdevhdr));
         data_dump(&cdevhdr, sizeof(cdevhdr));
     }
 
@@ -594,7 +594,7 @@ char            pathname[MAX_PATH];     /* file path in host format  */
     /* L1TAB itself is not adjusted for endian-ness                  */
     if (cmd_l1tab) {
         fprintf(stderr, "\nL1TAB - %d (0x%X) bytes:\n",
-                (ssize_t)(n * CCKD_L1ENT_SIZE), (ssize_t)(n * CCKD_L1ENT_SIZE));
+                (int)(n * CCKD_L1ENT_SIZE), (unsigned int)(n * CCKD_L1ENT_SIZE));
         data_dump(l1, n * CCKD_L1ENT_SIZE);
     }
 
@@ -656,11 +656,11 @@ char            pathname[MAX_PATH];     /* file path in host format  */
         if (cmd_l2tab) {
             fprintf(stderr, 
                    "\nL2TAB - %d (decimal) bytes\n", 
-                   (ssize_t)(cdevhdr.numl2tab * sizeof(CCKD_L2ENT)));
+                   (int)(cdevhdr.numl2tab * sizeof(CCKD_L2ENT)));
             data_dump(l2, (cdevhdr.numl2tab * sizeof(CCKD_L2ENT)) );
         }
         fprintf(stderr, "\nL2 index %d = L2TAB entry %d bytes\n",
-               l2ndx, (ssize_t)sizeof(CCKD_L2ENT) );
+               l2ndx, (int)sizeof(CCKD_L2ENT) );
         data_dump(&l2[l2ndx], sizeof(CCKD_L2ENT) );
         trkhdroff = l2[l2ndx].pos;
         imglen = l2[l2ndx].len;
